ENHANCE: Add toolbar and refine shortcuts

+++ heremes-data-tagger/Photo.cs
@@ -17,32 +17,108 @@ namespace HermesDataTagger
         public string Identifier { get; }
 
         #region TaggedItems
-        public BindingList<TaggedPerson> TaggedPeople = new BindingList<TaggedPerson>();
-        public TaggedPerson LastPersonTagged => TaggedPeople.FirstOrDefault();
-        public bool HasTaggedPeople => TaggedBibNumbers.Length > 0;
-        private string[] TaggedBibNumbers => TaggedPeople.Select(p => p.BibNumber).ToArray();
-        public TaggedPerson SelectedPerson { get; set; }
+        public event EventHandler SelectedRunnerUpdated;
+        public BindingList<TaggedPerson> TaggedRunners = new BindingList<TaggedPerson>();
+        public TaggedPerson LastRunnerTagged => TaggedRunners.FirstOrDefault();
+        public bool HasTaggedARunner => TaggedBibNumbers.Length > 0;
+        private string[] TaggedBibNumbers => TaggedRunners.Select(p => p.BibNumber).ToArray();
+        private TaggedPerson _selectedRunner;
+        public bool IsRunnerSelected => SelectedRunner != null;
+        public bool CanSelectNextRunner => HasTaggedARunner && SelectedRunner != TaggedRunners.Last();
+        public bool CanSelectPrevRunner => HasTaggedARunner && SelectedRunner != TaggedRunners.First();
+        public string SelectedRunnerNumber => IsRunnerSelected ? $"Selected Runner: {SelectedRunner.BibNumber}" : "";
+        public TaggedPerson SelectedRunner
+        {
+            get => _selectedRunner;
+            set { 
+                _selectedRunner = value;
+                if (SelectedRunnerUpdated != null && value != null)
+                {
+                    SelectedRunnerUpdated(this, EventArgs.Empty);
+                }
+            }
+        }
+        public void SelectNextRunner()
+        {
+            if (CanSelectNextRunner)
+            {
+                SelectedRunner = TaggedRunners[TaggedRunners.IndexOf(SelectedRunner) + 1];
+            }
+            else if (HasTaggedARunner)
+            {
+                SelectedRunner = TaggedRunners.First();
+            }
+        }
+        public void SelectPrevRunner()
+        {
+            if (CanSelectPrevRunner)
+            {
+                SelectedRunner = TaggedRunners[TaggedRunners.IndexOf(SelectedRunner) - 1];
+            }
+            else if (HasTaggedARunner)
+            {
+                SelectedRunner = TaggedRunners.Last();
+            }
+        }
+
+        public void DeleteTaggedPerson(TaggedPerson person)
+        {
+            TaggedRunners.Remove(person);
+            if (person == SelectedRunner)
+            {
+                SelectedRunner = null;
+            }
+            if (TaggedRunners.Count == 0)
+            {
+                TaggingStep = StepType.SelectBibRegion;
+            }
+        }
         #endregion
 
         #region Steps
         // Steps in tagging the photo
-        public StepType TaggingStep { get; set; }
+        private StepType _taggingStep = StepType.ImageCrowded;
+        public StepType TaggingStep {
+            get => _taggingStep;
+            set {
+                switch (value)
+                {
+                    case StepType.ImageCrowded:
+                        _taggingStep = value;
+                        break;
+                    case StepType.SelectBibRegion:
+                        if (!CanMarkBibs)
+                        {
+                            _taggingStep = StepType.ImageCrowded;
+                            MessageBox.Show("You cannot select bib regions as this image has been marked as crowded", "Note", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        }
+                        else
+                        {
+                            _taggingStep = value;
+                        }
+                        break;
+                    case StepType.SelectFaceRegion:
+                        if (!CanMarkFaces)
+                        {
+                            _taggingStep = CanMarkBibs ? StepType.SelectBibRegion : StepType.ImageCrowded;
+                            MessageBox.Show("You cannot tag image regions as there are no bib regions tagged", "Note", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        }
+                        else
+                        {
+                            _taggingStep = value;
+                        }
+                        break;
+                }
+            }
+        }
         public bool IsFirstTaggingStep => TaggingStep.IsFirstStep();
         public bool IsLastTaggingStep => TaggingStep.IsLastStep();
         public string TaggingStepName => TaggingStep.ToStepNameString();
         public string TaggingStepInstructions => TaggingStep.ToInstructionString();
+        public bool CanMarkFaces => CanMarkBibs && HasTaggedARunner;
+        public bool CanMarkBibs => IsPhotoNotCrowded;
         public void GoToNextStep()
         {
-            // Can only progress if photo is not crowded
-            if (IsPhotoCrowded)
-            {
-                return;
-            }
-            // Can only progress to following steps if at least one person tagged
-            if (TaggingStep == StepType.SelectBibRegion && !HasTaggedPeople)
-            {
-                return;
-            }
             if (TaggingStep != (StepType)(Enum.GetValues(typeof(StepType)).Length - 1))
             {
                 TaggingStep++;
@@ -59,9 +135,27 @@ namespace HermesDataTagger
 
         #region GeneralClassifications
         // General classifications about the photo
-        private bool _photoCrowded = false;
-        public bool IsPhotoNotCrowded { get { return !_photoCrowded; } }
-        public bool IsPhotoCrowded    { get { return _photoCrowded; }  set { _photoCrowded = value; } }
+        private bool _isPhotoCrowded = false;
+        public bool IsPhotoNotCrowded { get { return !IsPhotoCrowded; } set { IsPhotoCrowded = !value; } }
+        public bool IsPhotoCrowded
+        {
+            get => _isPhotoCrowded;
+            set
+            {
+                _isPhotoCrowded = value;
+                if (_isPhotoCrowded)
+                {
+                    TaggingStep = StepType.ImageCrowded;
+                    // TODO: Reset all people tagged?
+                    SelectedRunner = null;
+                    MainWindow.Singleton.RequestRedrawGraphics();
+                }
+                else if (!_isPhotoCrowded && TaggedRunners.Count > 0)
+                {
+                    SelectedRunner = TaggedRunners.First();
+                }
+            }
+        }
         #endregion
 
         public Photo(string filename)
@@ -71,14 +165,6 @@ namespace HermesDataTagger
             TaggingStep = StepType.ImageCrowded;
         }
 
-        public void DeleteTaggedPerson(TaggedPerson person)
-        {
-            TaggedPeople.Remove(person);
-            if (person == SelectedPerson)
-            {
-                SelectedPerson = null;
-            }
-        }
 
         #region HandleEvents
         public void HandleClick(PictureBox pbx, MouseEventArgs e)
@@ -106,6 +192,17 @@ namespace HermesDataTagger
                     break;
             }
         }
+
+        internal void RedoLastAction()
+        {
+            throw new NotImplementedException();
+        }
+
+        internal void UndoLastAction()
+        {
+            throw new NotImplementedException();
+        }
+
         public void HandleDragMove(PictureBox pbx, MouseEventArgs e)
         {
             switch (TaggingStep)
@@ -123,10 +220,10 @@ namespace HermesDataTagger
             {
                 case StepType.SelectFaceRegion:
                     UpdateEndOfFaceRegion(pbx, e.Location);
-                    bool didSetBothClassifications = AskForBaseClassificationsOfPerson(SelectedPerson) && AskForColorClassificationsOfPerson(SelectedPerson);
+                    bool didSetBothClassifications = AskForBaseClassificationsOfPerson(SelectedRunner) && AskForColorClassificationsOfPerson(SelectedRunner);
                     if (!didSetBothClassifications)
                     {
-                        SelectedPerson.Face.ClearPoints();
+                        SelectedRunner.Face.ClearPoints();
                     }
                     break;
                 default:
@@ -136,6 +233,10 @@ namespace HermesDataTagger
         #endregion
 
         #region CrowdedPhoto
+        public void ToggleCrowdedPhoto()
+        {
+            IsPhotoCrowded = !IsPhotoCrowded;
+        }
         public void AskIfPhotoCrowded()
         {
             DialogResult result = MessageBox.Show("Is this photo crowded?", "Crowded Image", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
@@ -152,28 +253,28 @@ namespace HermesDataTagger
                 return;
             }
 
-            TaggedPerson person = LastPersonTagged;
+            TaggedPerson person = LastRunnerTagged;
 
             // First person tagged or new person (last person has 4 clicks)?
             if (person == null || person.Bib.ClickPoints.Count == 4)
             {
                 person = new TaggedPerson(this);
-                TaggedPeople.Insert(0, person);
-                Debug.WriteLine($"Adding person #{TaggedPeople.Count} to ({Identifier})");
+                TaggedRunners.Insert(0, person);
+                Debug.WriteLine($"Adding person #{TaggedRunners.Count} to ({Identifier})");
             }
             person.Bib.ClickPoints.Add(pt);
             person.Bib.PixelPoints.Add(pt.ToPixelPoint(pbx));
-            Debug.WriteLine($"Person #{TaggedPeople.Count} has Bib[{person.Bib.ClickPoints.Count}] = {pt} ({Identifier})");
+            Debug.WriteLine($"Person #{TaggedRunners.Count} has Bib[{person.Bib.ClickPoints.Count}] = {pt} ({Identifier})");
             // We just finished tagging?
             if (person.IsBibRegionTagged)
             {
                 // Invalidate (update graphics) of picture box to reflect new bib number
                 pbx.Invalidate();
-                AskToTagBibNumber(pbx, person);
+                AskToTagBibNumber(pbx, person, true);
             }
         }
 
-        public void AskToTagBibNumber(PictureBox pbx, TaggedPerson person)
+        public void AskToTagBibNumber(PictureBox pbx, TaggedPerson person, bool shouldDeleteIfCancel = false)
         {
             // Can only tag if all clicked!
             if (person.IsBibRegionTagged)
@@ -183,14 +284,14 @@ namespace HermesDataTagger
                 do
                 {
                     DialogResult result = bibDiag.ShowDialog();
-                    if (result == DialogResult.Cancel)
+                    if (result == DialogResult.Cancel && shouldDeleteIfCancel)
                     {
                         // Cancel -- remove this tag!
                         DeleteTaggedPerson(person);
                         return;
                     }
                     string diagBibNumber = bibDiag.EnteredBibNumber;
-                    if (TaggedBibNumbers.Contains(diagBibNumber))
+                    if (TaggedBibNumbers.Contains(diagBibNumber) && person.BibNumber != diagBibNumber)
                     {
                         MessageBox.Show($"The bib number {diagBibNumber} already exists in this photo!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                     }
@@ -201,8 +302,9 @@ namespace HermesDataTagger
                     }
                 } while (true);
                 // Notify that bindings should be updated to display tag
-                TaggedPeople.ResetItem(0);
-                Debug.WriteLine($"Person #{TaggedPeople.Count} RBN identified as {person.Bib.BibNumber} ({Identifier})");
+                TaggedRunners.ResetItem(0);
+                MainWindow.Singleton.RequestRedrawGraphics();
+                Debug.WriteLine($"Person #{TaggedRunners.Count} RBN identified as {person.Bib.BibNumber} ({Identifier})");
             }
         }
         #endregion
@@ -210,19 +312,19 @@ namespace HermesDataTagger
         #region FaceRegion
         public void RecordStartOfFaceRegion(PictureBox pbx, Point pt)
         {
-            Debug.WriteLine($"Person #{SelectedPerson.BibNumber} face reigon start at {pt} ({Identifier})");
+            Debug.WriteLine($"Person #{SelectedRunner.BibNumber} face reigon start at {pt} ({Identifier})");
             SetFaceReigonAtIndex(pbx, pt, 0);
         }
 
         public void UpdateEndOfFaceRegion(PictureBox pbx, Point pt)
         {
-            Debug.WriteLine($"Person #{SelectedPerson.BibNumber} face reigon end at {pt} ({Identifier})");
+            Debug.WriteLine($"Person #{SelectedRunner.BibNumber} face reigon end at {pt} ({Identifier})");
             SetFaceReigonAtIndex(pbx, pt, 1);
         }
 
         private void SetFaceReigonAtIndex(PictureBox pbx, Point pt, int idx)
         {
-            TaggedPerson.PersonFace face = SelectedPerson.Face;
+            TaggedPerson.PersonFace face = SelectedRunner.Face;
             Point pixelPt = pt.ToPixelPoint(pbx);
 
             if (face.ClickPoints.Count == idx)
@@ -245,7 +347,7 @@ namespace HermesDataTagger
             bool didSet = dialog.ShowDialog() != DialogResult.Cancel;
             if (didSet)
             {
-                TaggedPeople.ResetItem(TaggedPeople.IndexOf(person));
+                TaggedRunners.ResetItem(TaggedRunners.IndexOf(person));
             }
             return didSet;
         }
@@ -255,7 +357,7 @@ namespace HermesDataTagger
             bool didSet = dialog.ShowDialog() != DialogResult.Cancel;
             if (didSet)
             {
-                TaggedPeople.ResetItem(TaggedPeople.IndexOf(person));
+                TaggedRunners.ResetItem(TaggedRunners.IndexOf(person));
             }
             return didSet;            
         }
